Criando os metodos para cadastrar um novo cliente
