+++ projetopousada/src/negocios/NegocioCliente.java
@@ -8,18 +8,13 @@ import negocios.Pessoa.*;
 
 public class NegocioCliente {
 	
-	private IRepositorioPessoa pessoas;
 	private IRepositorioQuarto quartos;
+	private IRepositorioPessoa pessoas;
 	
-	public NegocioCliente() {
-		this.pessoas = new RepositorioPessoaArrayList();
-		this.quartos = new RepositorioQuartoArrayList();
+	public NegocioCliente(IRepositorioPessoa repositorioPessoa, IRepositorioQuarto repositorioQuarto) {
+		this.pessoas = repositorioPessoa;
+		this.quartos = repositorioQuarto;
 	}
 	
-	public void cadastrarNovoCliente(Gerente adm, String nome, String cpf, String senha) {
-		Cliente cliente = new Cliente(nome, cpf, senha);
-		//tratar a exception de addPessoa
-		adm.addPessoa(pessoas, cliente);
-	}
 	
 }
