+++ projetopousada/src/negocios/Pessoa/Gerente.java
@@ -26,33 +26,6 @@ public class Gerente extends Pessoa  {
 			pessoas.addPessoa(pessoa);
 			return 1;
 		}
-		
-		/*
-		String nome,cpf;
-		String tipoUsario;
-		if (tipoPessoa==1) {
-			tipoUsario = "cliente";
-		}
-		else {
-			tipoUsario = "gerente";
-		}
-		String senha;
-		int busca;
-		Scanner scan = new Scanner(System.in);
-		do {
-			System.out.println("Digite o CPF do novo "+tipoUsario +":");
-	        cpf = scan.nextLine();
-	        busca = pessoas.buscarPessoa(cpf);
-	        if (busca!=-1) {
-	        	System.out.println("operacao invalida, CPF cadastrado anteriormente");
-	        }
-		}while(busca!=-1);
-		System.out.println("Qual nome do novo "+tipoUsario+":");
-        nome = scan.nextLine();
-		System.out.println("Qual a senha do novo "+tipoUsario+":");
-		senha = scan.nextLine();
-		pessoas.criarPessoa(nome, cpf, senha, tipoPessoa);
-		*/
 	}
 	//Lembrar adicionar a excetion do Repositorio Lotado
 	public int addPessoa(IRepositorioPessoa pessoas, int tipoPessoa, String nome, 
@@ -77,39 +50,6 @@ public class Gerente extends Pessoa  {
 		else {
 			return 1;
 		}
-		
-		/*
-		String cpf;
-		int busca, valido = -1;
-		Scanner scan = new Scanner(System.in);
-        do {
-        	System.out.println("Digite o CPF da Pessoa:");
-            cpf = scan.nextLine();
-            busca = pessoas.buscarPessoa(cpf);
-            if(busca == -1) {
-            	valido = -1;
-            	System.out.println("operacao invalida, CPF nao cadastrado");
-            }
-            else if(busca == indexGerente) {
-            	valido = -1;
-            	System.out.println("CPF nao pode ser excluido");
-            }
-            else {
-            	valido = 1;
-            }
-            	
-        }while(valido == -1);
-        System.out.println(pessoas.getPessoas().get(busca).toString());
-        System.out.println("Confirma remocao (0 - nao, 1 - sim): ");
-        int escolha = scan.nextInt();
-        if (escolha == 1) {
-        	pessoas.deletarPessoa(cpf);
-        	System.out.println("remocao conluida");
-        }
-        else {
-        	System.out.println("remocao cancelada");
-        }
-        */
 	}
 	public int removerPessoa(IRepositorioPessoa pessoas, String cpf) {
 		int result = pessoas.deletarPessoa(cpf);
@@ -177,21 +117,6 @@ public class Gerente extends Pessoa  {
 			}
 			return 1;
 		}
-		
-		/*
-		int numQuarto;
-		Scanner scan = new Scanner(System.in);
-		int busca;
-		do {
-			System.out.println("qual o numero do quarto?");
-			numQuarto= scan.nextInt();
-			busca = quartos.buscarQuarto(numQuarto, tipoQuarto);
-			if (busca!=-1) {
-	        	System.out.println("operacao invalida, numero do quarto cadastrado anteriormente");
-	        }
-		}while(busca!=-1);
-		quartos.adicionarQuarto(numQuarto, tipoQuarto);
-		*/
 	}
 	public int addQuarto(IRepositorioQuarto quartos, int numeroQuarto, int tipoQuarto) {
 		int indexQuarto = quartos.buscarQuarto(numeroQuarto, tipoQuarto);
@@ -219,30 +144,6 @@ public class Gerente extends Pessoa  {
 		else {
 			return 1;
 		}
-		
-		
-		/*
-		int numQuarto, busca;
-		Scanner scan = new Scanner(System.in);
-		do {
-			System.out.println("qual o numero do quarto?:");
-			numQuarto= scan.nextInt();
-			busca = quartos.buscarQuarto(numQuarto, tipoQuarto);
-			if(busca == -1) {
-            	System.out.println("operacao invalida, Quarto nao cadastrado");
-            }
-		}while(busca==-1);
-		System.out.println(quartos.getQuarto(busca).toString());
-		System.out.println("Confirma remocao (0 - nao, 1 - sim): ");
-        int escolha = scan.nextInt();
-        if (escolha == 1) {
-        	quartos.deletarQuarto(numQuarto, tipoQuarto);
-        	System.out.println("remocao conluida");
-        }
-        else {
-        	System.out.println("remocao cancelada");
-        }
-        */
 	}
 	public int deletarQuarto(IRepositorioQuarto quartos, int numeroQuarto, int tipoQuarto) {
 		int result = quartos.deletarQuarto(numeroQuarto, tipoQuarto);
